requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
astroid==2.3.2
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
Click==7.0
coverage==4.5.4
cryptography==3.2.1
distlib==0.3.1
factory-boy==2.12.0
Faker==2.0.3
filelock==3.0.12
Flask==1.1.1
Flask-Bcrypt==0.7.1
flask-restplus==0.13.0
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
idna==2.10
importlib-metadata==0.23
install==1.3.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
protobuf==3.14.0
psycopg2-binary==2.8.3
py==1.8.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.4.3
pymongo==3.11.2
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.2.1
pytest-cov==2.8.1
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-html==2.0.0
pytest-metadata==1.8.0
pytest-sugar==0.9.2
python-dateutil==2.8.0
pytz==2019.3
requests==2.25.1
six==1.12.0
SQLAlchemy==1.3.10
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.2
virtualenv==20.3.0
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
